+++ src/test/java/com/metrica/formacion/TestService.java
@@ -1,15 +1,12 @@
 package com.metrica.formacion;
 
 
-import com.metrica.formacion.config.SpringConfigurationFile;
 import com.metrica.formacion.dao.GrupoRepository;
 import com.metrica.formacion.entity.grupos;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import java.time.LocalTime;
