+++ src/test/java/com/metrica/formacion/dao/GrupoRepositoryTest.java
@@ -1,10 +1,8 @@
 package com.metrica.formacion.dao;
 
-import com.metrica.formacion.config.SpringConfigurationFile;
+import com.metrica.formacion.SpringConfigurationFile;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest(classes = SpringConfigurationFile.class)
 @Transactional
